Home 🏠

Functions addResourcePath tablerApp row column col12 col11 col10 col9 col8 col7 col6 col5 col4 col3 col2 col1 card pre_title title header body value_box icon alert button page topbar footer sidebar_menu sidebar_brand horizontal_menu menu_item menu_dropdown tab_items tab_item navbar_menu tagAppendAttributes tagList HTML div span p a img ul li aside h1 h2 h3 h4 h5 h6 hr br selectInput selectMultipleInput sliderInput updateSelectInput updateSliderInput textInput numericInput dateInput checkboxInput actionButton checkboxGroupInput radioButtons NS getDefaultReactiveDomain moduleServer textOutput verbatimTextOutput uiOutput renderText renderPrint renderUI plotOutput renderPlot widgetOutput renderWidget downloadHandler downloadButton downloadLink showProgress hideProgress withProgress reactiveVal reactiveValues reactiveValuesToList reactive req isolate observe observeEvent eventReactive bindEvent tablerOptions bindCache syncUrl pkg_template show hide toggle hidden

News Changelog

visibilityFuncs {tabler}	R Documentation

Show/Hide an element

Description

Show or hide an HTML element in a running tablerApp.

show makes an element visible, hide makes an element invisible, toggle displays the element if it is hidden and hides it if it is visible.

showElement, hideElement, and toggleElement are synonyms for show/hide/ toggle.

If condition is given to toggle, that condition is used to decide whether to show or hide the element: the element is shown when the condition evaluates to TRUE and hidden otherwise.

Usage

show(
  session = getDefaultReactiveDomain(),
  id = NULL,
  anim = FALSE,
  animType = "slide",
  time = 0.5,
  selector = NULL
)

showElement(
  session = getDefaultReactiveDomain(),
  id = NULL,
  anim = FALSE,
  animType = "slide",
  time = 0.5,
  selector = NULL
)

hide(
  session = getDefaultReactiveDomain(),
  id = NULL,
  anim = FALSE,
  animType = "slide",
  time = 0.5,
  selector = NULL
)

hideElement(
  session = getDefaultReactiveDomain(),
  id = NULL,
  anim = FALSE,
  animType = "slide",
  time = 0.5,
  selector = NULL
)

toggle(
  session = getDefaultReactiveDomain(),
  id = NULL,
  anim = FALSE,
  animType = "slide",
  time = 0.5,
  selector = NULL,
  condition = NULL
)

toggleElement(
  session = getDefaultReactiveDomain(),
  id = NULL,
  anim = FALSE,
  animType = "slide",
  time = 0.5,
  selector = NULL,
  condition = NULL
)

Arguments

session 	 The session object passed by tablerApp to the server function.

id 	 The id of the element/HTML tag.

anim 	 If TRUE then animate the behaviour.

animType 	 The type of animation to use, either "slide" or "fade".

time 	 The number of seconds to make the animation last.

selector 	 CSS selector of the elements to show/hide. Ignored if the id argument is given. For example, to select all span elements with class x, use selector = "span.x".

condition 	 An optional argument to toggle, see 'Details' below.

Details

This is a dependency-free adaptation of shinyjs's show/hide/toggle functions for tabler, which does not use or depend on shiny. Instead of relying on a Shiny session, these functions use the plain-list session object created by tablerApp and passed as the third argument to the server function.

See Also

tablerApp

Examples

if (interactive()) {
  ui <- page(
    title = "Show/Hide Example",
    body = body(
      actionButton("btn", "Click me"),
      div(id = "panel", "Watch what happens to me")
    )
  )

  server <- function(input, output, session) {
    observeEvent(input$btn, {
      # Change the following line for more examples
      toggle(session, "panel")
    })
  }

  tablerApp(ui, server)
}
## Not run:
# The function call in the above app can be replaced by any of the
# following examples to produce similar apps
toggle(session, id = "panel")
toggle(session, "panel", TRUE)
toggle(session, "panel", TRUE, "fade", 2)
toggle(session, id = "panel", time = 1, anim = TRUE, animType = "slide")
show(session, "panel")
show(session, id = "panel", anim = TRUE)
hide(session, "panel")
hide(session, id = "panel", anim = TRUE)

## End(Not run)

## toggle can be given an optional `condition` argument, which
## determines if to show or hide the element
if (interactive()) {
  ui <- page(
    title = "Conditional Toggle",
    body = body(
      checkboxInput("checkbox", "Show the text", TRUE),
      div(id = "element", "Watch what happens to me")
    )
  )

  server <- function(input, output, session) {
    observe({
      toggle(session, id = "element", condition = input$checkbox)
    })
  }

  tablerApp(ui, server)
}